app/requirements.txt
fastapi>=0.111.0
google-generativeai>=0.7.1
huggingface-hub>=0.23.4
langchain>=0.2.1
langchain-community>=0.2.1
langchain-core>=0.2.1
langchain-google-genai>=1.0.6
langchain-chroma>=0.1.2
chromadb>=0.5.3
pypdf>=4.2.0
python-dotenv>=1.0.1
redis>=5.0.7
scikit-learn>=1.5.0
sentence-transformers>=3.0.1
tqdm>=4.66.4
uvicorn>=0.30.1